ACCUMULO-804 avoid spewing warnings when deleting missing WALogs, which is the normal case
git-svn-id: https://svn.apache.org/repos/asf/accumulo/branches/1.5@1458981 13f79535-47bb-0310-9956-ffa450edef68

+++ server/src/main/java/org/apache/accumulo/server/gc/GarbageCollectWriteAheadLogs.java
@@ -189,13 +189,13 @@ public class GarbageCollectWriteAheadLogs {
       log.debug("Removing sorted WAL " + sortedWALog);
       try {
         Path swalog = new Path(recoveryDir, sortedWALog);
-        if (trash == null || !trash.moveToTrash(swalog)) {
+        if (trash == null || (fs.exists(swalog) && !trash.moveToTrash(swalog))) {
           fs.delete(swalog, true);
         }
       } catch (FileNotFoundException ex) {
         // ignored
       } catch (IOException ioe) {
-        log.error("Unable to delete sorted walog " + sortedWALogs + ": " + ioe);
+        log.error("Unable to delete sorted walog " + sortedWALog + ": " + ioe);
       }
     }
